netlink: introduce UpdateDeviceConfig
This new method takes care of setting up the IP address and routes of a
network device based on configuration received from a wiresteward
server. Some of the functionality previously part of the Agent struct
has been replaced by this change.

Additionally, the darwin implementation should be more efficient, using
the minimum amount of sockets to configure a device.

Finally, a new struct PeerConfig is being introduced which holds
the parsed wiresteward configuration, in a form suitable for device
configuration.

Overall, this is an intermediate commit towards refining the structs.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>