Introduce DeviceManager
The DeviceManager wraps around a TunDevice and implements all the
operation relating to the configuration of the device based on
information received from a wiresteward server. This includes both the
device address, routing table routes and wireguard peers.

A lot of this functionality has been migrated from the Agent struct.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>