TunDevice is reusable
Previously a TunDevice was initialised on creation and Run() could be
called after to start processing events, with Stop() stopping all
underlying processes.

This change moves initialiation inside Run() which does not wait and
instead spawn a goroutine internally. Stop() can be used to stop the
device as before but also, the device can now be re-started by calling
Run() again.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>