Bring the interface up before setting an address
It appears that the interface comes "up" automatically a while after
being created and regardless of when LinkSetUp() is called (which does
not wait for it to come up). Since the `OperState` field eventually
comes up as "unknown", we don't want to rely on its value and instead
check for the presence of the up flag on the device.

If we call AddrAdd() before the interface comes up, netlink seems to
register the address (which can be confirmed by AddrList()) but the
address does not appear on the system (for example with ip a). Waiting
for the condition described above addresses the issue sufficiently.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>