Call LinkSetUp after timeout
In 97d8804d6d22b19f6213b1e30d787aa8b0521882 waitLinkIsUp was introduced
to address an issue with the link automatically transitioning to the UP
state. The issue described there was originally observed on flatcar
linux.

After further investigation, it appears that the behaviour described in
the commit above is not consistent across different systems. In Debian,
for example, after LinkAdd is called the link will remain in the DOWN
state and waitLinkIsUp will timeout.

Calling LinkSetUp before waitLinkIsUp will work in Debian but in flatcar
it creates a scenario where the link has the UP flag set but AddrAdd
still fails.

Instead, we can resort to calling LinkSetUp only when we timeout waiting
for the link to come up automatically.

This diverges from the order in which wg-quick sets up things but there
shouldn't be any complications arising from bringing the link up earlier
in the process and setting it up afterwards.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>