Keep single old config instead of list
Device manager will now connect to only one peer and needs to keep a single
configuration to be able to clean routes and interface ips on lease renewal.
This can remedy the buggy behaviour where old ips remained on interfaces as
dm.config[serverURL] was nil for the new healthy server and cleanup was not
triggered.