tf/matchbox: move DNS configuration to resolved.conf
The latest releases of Flatcar stable enable the systemd-resolved stub
resolver. Although I'm not sure why, this seems to ignore the DNS server
set on the link (even though systemd-resolved does properly populate
resolv.conf with it).

This commit sets the DNS server globally in /etc/systemd/resolved.conf.
It also adds the other Cloudflare DNS address 1.0.0.1 which we're
providing via DHCP on other instances.

I've made the nameservers configurable too, to avoid having to edit the
module if we want to change this in future.